requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
billiard==4.1.0
cachetools==5.3.1
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.5
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.7
google-auth==2.23.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.2
oauthlib==3.2.2
packaging==23.1
prompt-toolkit==3.0.39
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.3
rsa==4.9
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
